Refactored to use Svelte framework to modify state, switch webpack with vite
