--- .gitignore
@@ -1,3 +1,0 @@
-.DS_Store
-node_modules
-public/build/bundle.*
