+++ package.json
@@ -1,26 +1,22 @@
 {
-  "name": "svelte-app",
-  "version": "1.0.0",
-  "devDependencies": {
-    "@tsconfig/svelte": "^1.0.10",
-    "@types/node": "^14.11.1",
-    "cross-env": "^7.0.3",
-    "css-loader": "^5.0.1",
-    "mini-css-extract-plugin": "^1.3.4",
-    "svelte": "^3.31.2",
-    "svelte-check": "^1.0.46",
-    "svelte-loader": "^3.0.0",
-    "svelte-preprocess": "^4.3.0",
-    "ts-loader": "^8.0.4",
-    "tslib": "^2.0.1",
-    "typescript": "^4.0.3",
-    "webpack": "^5.16.0",
-    "webpack-cli": "^4.4.0",
-    "webpack-dev-server": "^3.11.2"
-  },
+  "name": "canvas-physics",
+  "private": true,
+  "version": "0.0.0",
+  "type": "module",
   "scripts": {
-    "build": "cross-env NODE_ENV=production webpack",
-    "dev": "webpack serve --content-base public",
-    "validate": "svelte-check"
+    "dev": "vite",
+    "build": "vite build",
+    "preview": "vite preview",
+    "check": "svelte-check --tsconfig ./tsconfig.json"
+  },
+  "devDependencies": {
+    "@sveltejs/vite-plugin-svelte": "^1.0.2",
+    "@tsconfig/svelte": "^3.0.0",
+    "svelte": "^3.49.0",
+    "svelte-check": "^2.8.1",
+    "svelte-preprocess": "^4.10.7",
+    "tslib": "^2.4.0",
+    "typescript": "^4.6.4",
+    "vite": "^3.1.0"
   }
 }
