+++ src/lib/Slider.svelte
