+++ src/lib/physics.ts
