+++ src/main.ts
@@ -1,7 +1,7 @@
-import App from './App.svelte';
+import App from './App.svelte'
 
 const app = new App({
-	target: document.body,
-});
+  target: document.getElementById('app')
+})
 
-export default app;
+export default app
